[Database Update] Added Recordings Model and Improved Prisma Client Initialization
The Prisma client initialization process in pages/libs/server/client.ts is modified. Now, a global variable client is declared, which can either be an existing PrismaClient instance or a new one. This ensures the reuse of the PrismaClient instance between requests in a serverless environment. Additionally, in development mode, the Prisma client is attached to the global object to allow its access across the application.

A new model Recordings is added to the Prisma schema in prisma/schema.prisma. The Recordings model includes fields such as id, createdAt, updatedAt, user, userId, image, name, and description. A relation is established between the User model and Recordings model with the userId field, with the onDelete policy set to Cascade. This model also includes an index on the userId field for efficient querying.